+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Skip query for buy together when `productIds` is empty.
+
 ## [1.1.0] - 2020-12-09
 
 ### Changed